+++ p2p/src/protocol/mod.rs
@@ -1,2 +1,3 @@
+pub mod address_sanity;
 pub mod messages;
 pub mod peer;